Started fixing broken links
I did this because the file management system was messy and horrible to use. I plan on implementing JS and more uniform and straigtforward ways of referencing documents, so that there is not so much that needs to be changed in every document, and so there is less repetitive code.